Move all non-cmd code to the pkg package
