+++ cmd/up.go
@@ -5,13 +5,13 @@ import (
 
 	log "github.com/sirupsen/logrus"
 
-	"github.com/okteto/cnd/k8/client"
-	"github.com/okteto/cnd/k8/deployments"
-	"github.com/okteto/cnd/k8/forward"
-	"github.com/okteto/cnd/storage"
-	"github.com/okteto/cnd/syncthing"
+	"github.com/okteto/cnd/pkg/k8/client"
+	"github.com/okteto/cnd/pkg/k8/deployments"
+	"github.com/okteto/cnd/pkg/k8/forward"
+	"github.com/okteto/cnd/pkg/storage"
+	"github.com/okteto/cnd/pkg/syncthing"
 
-	"github.com/okteto/cnd/model"
+	"github.com/okteto/cnd/pkg/model"
 	"github.com/spf13/cobra"
 )
