+++ pkg/syncthing/configXML.go
