+++ pkg/syncthing/syncthing.go
